BUGFIX: Dynamic lattice allocation had broken 'Z' interaction.
This was due to the (pre processor) directive to choose a 2D lattice
being selected, even when running in 3D.
As a result, only 2D slices within Z were considered.
Similar 2D patterns were the result (naturally!), but this caused the 3D
correlation functions (RDF+ etc.) to be incorrect, as there was no
correlation in the Z direction between the layers.